+++ src/Runtime/Matcher/AnyChildMatcher.php
@@ -3,12 +3,12 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Path\Runtime\Matcher;
 
-use Remorhaz\JSON\Data\Value\ValueInterface;
+use Remorhaz\JSON\Data\Value\NodeValueInterface;
 
 final class AnyChildMatcher implements ChildMatcherInterface
 {
 
-    public function match($address, ValueInterface $value): bool
+    public function match($address, NodeValueInterface $value, NodeValueInterface $container): bool
     {
         return true;
     }