+++ src/Runtime/Matcher/ChildMatcherInterface.php
@@ -3,10 +3,10 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Path\Runtime\Matcher;
 
-use Remorhaz\JSON\Data\Value\ValueInterface;
+use Remorhaz\JSON\Data\Value\NodeValueInterface;
 
 interface ChildMatcherInterface
 {
 
-    public function match($address, ValueInterface $value): bool;
+    public function match($address, NodeValueInterface $value, NodeValueInterface $container): bool;
 }