+++ src/Runtime/Matcher/StrictElementMatcher.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace Remorhaz\JSON\Path\Runtime\Matcher;
 
 use function in_array;
-use Remorhaz\JSON\Data\Value\ValueInterface;
+use Remorhaz\JSON\Data\Value\NodeValueInterface;
 
 final class StrictElementMatcher implements ChildMatcherInterface
 {
@@ -16,7 +16,7 @@ final class StrictElementMatcher implements ChildMatcherInterface
         $this->indice = $indice;
     }
 
-    public function match($address, ValueInterface $value): bool
+    public function match($address, NodeValueInterface $value, NodeValueInterface $container): bool
     {
         return in_array($address, $this->indice, true);
     }